Improved sprite boundles and added action ui

+++ src/main/java/com/maanoo/tredory/engine/Sprite.java
@@ -19,7 +19,7 @@ public abstract class Sprite {
         draw(x, y, width, height, Color.white);
     }
 
-    public void draw(float x, float y, float scale) {
+    public void draw(float x, float y, int scale) {
         draw(x, y, getImage().getWidth() * scale, getImage().getHeight() * scale);
     }
